abstracted card to be its own component, added all driver info into drivercards
