+++ frontend/src/components/Card/card.module.css
@@ -46,7 +46,7 @@
   margin-right: 16px;
 }
 
-.info {
+.infoContainer p {
   font-size: 15px;
   opacity: 0.7;
 }
